We will now listen for resize events in a singleton handle for all imgix images This will ensure only 1 resize listeners is added to the window for all imcix images on the page, when having 1k images, this is much faster
* Removes 'ember-resize-mixin' from package.json and its references in 'imgix-image-element' and 'imgix-image' components
* Removes '_onResize' handler from 'imgix-path-behavior' in favor of subscribing to the top level 'imgix-resize-listener' singleton
* Moves the debounce rate handling into the init of the new singleton, this is much faster than calculating it on every resize event

The reasoning behind this change is we were constantly seeing errors from the resize mixins and it seems like Ember has no elegant way to handle this.
Typically in single page apps (especially in hybrid mobile land) you want to limit the number of handlers you attach to the window, in favor of a few top level handlers that delegate inwards.
I believe this to be a better more performant approach.